Int test restructure
