+++ tests/integration/.gitignore
@@ -5,5 +5,5 @@ junit-report.xml
 junit-main-report.xml
 junit-upgrade-report.xml
 reports
-manifests/generated-operator-manifest.yaml
+steps/operator_generated_manifest.yaml
