+++ tests/integration/pkg/crds/crd_lister.go
@@ -1,4 +1,4 @@
-package manifests
+package crds
 
 import (
 	"context"
