+++ tests/integration/steps/istio_cr_ext_authz_template.yaml
