+++ tests/integration/steps/istio_cr_template.yaml
