+++ tests/integration/steps/operator_failing_manifest.yaml
